feat: migrate unit tests to path attributes for first 4 modules
- Migrate exit_codes.rs tests (~35 lines) to src/unit_tests/exit_codes_tests.rs
- Migrate main.rs tests (~66 lines) to src/unit_tests/main_tests.rs
- Migrate logging.rs tests (~71 lines) to src/unit_tests/logging_tests.rs
- Migrate hooks.rs tests (~226 lines) to src/unit_tests/hooks_tests.rs

Replace inline #[cfg(test)] mod tests blocks with path attributes:
#[cfg(test)]
#[path = "unit_tests/{module}_tests.rs"]
mod tests;

This improves source file readability by removing ~398 lines of test code
from implementation files and enables faster compilation by avoiding test
parsing during regular builds.

All tests pass with no functionality changes.
